+++ changelog.md
@@ -1,4 +1,5 @@
 # ACEmulator Change Log
+
 ### 2017-05-03
 [Jyrus]
 * Changed the structure of the table that contains the Portal Destinations to include other portal properties
@@ -13,6 +14,7 @@
 
 ### 2017-05-02
 [fantoms]
+* Built out the `@home` admin command.
 * Built out the `@save` functionality.
 * Added useful position debug commands, `@teletype`, `@setposition`, and `@listpositions`. `@teletype` will teleport the player, `@setposition` will test the database save functionality, and `@listpositions` will print out a list of all database positions too chat.
 * Updated `SaveCharacterPosition` too prevent PortalType.Undef types from being used as a valid position type.